better patch for POSIXlt changes in R-devel
Previous commit utilised a new function that was not in released R
so could not be sent to CRAN. Hacky approach for the time being
is to problematic attributes which also seems to remove the
balance_lt attribute.

Will revisit this in refactor where we should be able to avoid
any hackiness.