feat: (re)introduce grates_int_period objects
The 1.0.0 release of grates removed support for int_period
objects to narrow the packages scope to focuse on grouped
dates. Recently I was made aware that some users of the
downstream incidence2 package had been utilising said
functionality so I've decided to reintroduce
(see https://github.com/reconverse/incidence2/issues/121).

This undoes that change with an implementation somewhat
akin to grates_month objects.

I'm not publicising the functionality in the vignette
for the moment and it should be treated as "experimental"
until it gets wider usage.