feat: use better default for offset argument (breaking)
The default `offset` argument for the Date and POSIXt methods of the
`as_period()` generic is now set as the minimum of the input argument.
For character and factor methods, the default has been dropped and
must now be specified explicitly. Old behaviour for each method can
be obtained by setting `offset = 0`. Motivated by discussion with
@joshwlambert.

BREAKING CHANGE